requirements.txt
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
numpy
opencv-python
pillow
pyyaml
soundfile
scikit-learn
tqdm
matplotlib
